+++ src/concepts/Budget/views/BudgetBox/BudgetBox.module.scss
@@ -7,7 +7,7 @@
 .box {
   padding: 12px 0;
   background-color: var(--background-floor-1);
-  btransaction-radius: 5px;
+  border-radius: 5px;
   cursor: pointer;
 }
