+++ src/concepts/Budget/views/BudgetBox/hooks/useDeleteBudget.ts
@@ -1,6 +1,6 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { deleteBudgetRequest } from "@/concepts/infra/http/deleteBudgetRequest";
+import { deleteBudgetRequest } from "@/concepts/Budget/repository/deleteBudgetRequest";
 import { toastError } from "@/concepts/utils/toastify";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 
 export const useDeleteBudget = ({onSuccess}: any) => {
   const queryClient = useQueryClient()
@@ -12,6 +12,7 @@ export const useDeleteBudget = ({onSuccess}: any) => {
       queryClient.invalidateQueries({ queryKey: ['budgets'] })
     },
     onError: (error) => {
+      // @ts-ignore
       const errorMsg = error?.response?.data?.error?.details.join('. ') || 'Erro ao remover orçamento'
       toastError(errorMsg)
     }
