+++ src/concepts/transactions/index.ts
@@ -1,4 +1,2 @@
 export { TransactionProvider } from "./view/components/TransactionProvider";
-export { BudgetProvider } from "./view/components/BudgetProvider";
 export * from "./view/hooks/useTransaction";
-export * from "./view/hooks/useBudget";
