+++ src/concepts/transactions/view/hooks/useDeleteTransaction.ts
@@ -1,6 +1,6 @@
 
 
-import { useBudget } from "./useBudget";
+import { useBudget } from "../../../Budget/hooks/useBudget";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useTransaction } from "./useTransaction";
 import { formatDate } from "@/concepts/utils/date";
