+++ src/concepts/transactions/view/hooks/usePayTransaction.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query"
-import { useBudget } from "./useBudget";
+import { useBudget } from "../../../Budget/hooks/useBudget";
 import { useTransaction } from "./useTransaction";
 import { formatDate } from "@/concepts/utils/date";
 import { toastError } from "@/concepts/utils/toastify";
