+++ src/concepts/transactions/view/hooks/useUnpayTransaction.ts
@@ -1,6 +1,6 @@
 import { api } from "@/lib/http/api";
 import { useMutation, useQueryClient } from "@tanstack/react-query"
-import { useBudget } from "./useBudget";
+import { useBudget } from "../../../Budget/hooks/useBudget";
 import { useTransaction } from "./useTransaction";
 import { formatDate } from "@/concepts/utils/date";
 import { toastError } from "@/concepts/utils/toastify";
