+++ src/pages/budgets.tsx
@@ -1,5 +1,5 @@
 import { withSSRAuth } from "@/concepts/Auth/middlewares/withSSRAuth";
-import { BudgetsPage } from "@/concepts/pages/budgets/BudgetsPage";
+import { BudgetsPage } from "@/concepts/Budget/views/BudgetsPage";
 
 const Page = () => {
   return <BudgetsPage />;
